Cleanup, docs
